# docs/openapi.yaml
openapi: 3.0.3
info:
  title: InvestIQ API
  description: |
    InvestIQ is a comprehensive algorithmic trading platform with multi-engine analysis,
    backtesting, portfolio management, and autonomous trading capabilities.

    ## Authentication
    All endpoints (except `/health` and `/metrics`) require authentication via the `X-API-Key` header or Bearer token.

    ## Live Trading Protection
    Broker write endpoints (POST /api/broker/execute, DELETE /api/broker/positions/:symbol, POST /api/broker/orders/:id/cancel)
    require an additional `X-Live-Trading-Key` header for safety.

    ## Response Format
    All API responses follow a standard wrapper:
    ```json
    {
      "success": true,
      "data": { ... },
      "error": null
    }
    ```
  version: 1.0.0
  contact:
    name: InvestIQ Support
    url: https://github.com/yourusername/invest-iq
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.investiq.io
    description: Production server
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/agent/trades:
    get:
      tags: [Agent Trades]
      summary: List pending agent trade proposals
      responses:
        '200':
          description: List of pending trades
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags: [Agent Trades]
      summary: Agent proposes a new trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProposeTrade'
      responses:
        '200':
          description: Trade proposed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/agent/trades/{id}:
    get:
      tags: [Agent Trades]
      summary: Get a pending trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Pending trade details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/agent/trades/{id}/review:
    post:
      tags: [Agent Trades]
      summary: Approve or reject an agent trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [approve, reject]
      responses:
        '200':
          description: Trade reviewed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # ============================================================================
  # Alpha Decay Monitoring
  # ============================================================================
...
  /api/strategies/{name}/history:
    get:
      tags: [Alpha Decay]
      summary: Get performance history for a strategy
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Performance snapshots
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/strategies/{name}/decay:
    get:
      tags: [Alpha Decay]
      summary: Get decay analysis with change detection
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Decay analysis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/strategies/{name}/report:
    get:
      tags: [Alpha Decay]
      summary: Get full health report
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Health report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/strategies/snapshot:
    post:
      tags: [Alpha Decay]
      summary: Record a performance snapshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [strategy_name, snapshot_date, rolling_sharpe, win_rate, profit_factor, trades_count]
              properties:
                strategy_name:
                  type: string
                snapshot_date:
                  type: string
                  format: date
                rolling_sharpe:
                  type: number
                win_rate:
                  type: number
                profit_factor:
                  type: number
                trades_count:
                  type: integer
      responses:
        '200':
          description: Snapshot recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/strategies/{name}/retire:
    post:
      tags: [Alpha Decay]
      summary: Mark a strategy as retired
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Strategy retired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # ============================================================================
  # Flow Map (Sector Rotation)
  # ============================================================================
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication. Can also use Bearer token in Authorization header.

    LiveTradingAuth:
      type: apiKey
      in: header
      name: X-Live-Trading-Key
      description: Additional key required for live trading write operations.

    AdminAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Admin-level API key required for privileged operations.
  schemas:
    ApiResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        data:
          type: object
          description: Response data (structure varies by endpoint)
        error:
          type: string
          nullable: true
...
    ProposeTrade:
      type: object
      required: [symbol, action, shares]
      properties:
        symbol:
          type: string
        action:
          type: string
          enum: [buy, sell]
        shares:
          type: number
        confidence:
          type: number
        reason:
          type: string
        signal_type:
          type: string